Add multiple stages to the snabble extension and read them out
