+++ kotlin-customization-sample/build.gradle
@@ -33,7 +33,7 @@ android {
     }
 }
 
-snabble {
+snabble.production {
     appId = "snabble-sdk-demo-app-oguh3x"
     secret = "2TKKEG5KXWY6DFOGTZKDUIBTNIRVCYKFZBY32FFRUUWIUAFEIBHQ===="
     prefetchMetaData = true
