+++ ui/build.gradle
@@ -54,7 +54,6 @@ dependencies {
     implementation 'androidx.webkit:webkit:1.4.0'
     implementation 'androidx.viewpager2:viewpager2:1.0.0'
     implementation 'com.google.android.material:material:1.6.0'
-    implementation 'com.github.google:flexbox-layout:2.0.1'
     implementation 'me.relex:circleindicator:2.1.6'
     implementation 'ch.datatrans:android-sdk:1.5.0'
     implementation 'com.google.android.gms:play-services-wallet:19.1.0'
